game/requirements.txt
paho-mqtt==1.6.1
pymongo==4.6.1
cachetools==5.5.2 
pydantic==2.10.6
